+++ CHANGELOG.md
@@ -11,6 +11,8 @@ backwards compatibility.
 * Change contour range to 0.0 to 0.4 seconds linear up to 12 o'clock and
   0.4 to 10 logarithmic above that.
 * Rework the interface to provide more direct control over attributes.
+* Make it possible to offset scale steps by up to 4 quarter notes in either
+  direction.
 
 ## 0.1.1